4.21 done

+++ part4/controllers/blogs.js
@@ -12,18 +12,18 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.get('/:id', async (request, response, next) => {
-  const blog = await Blog.findById(request.params.id);
+  const blog = await Blog.findById(request.params.id)
 
   if (blog) {
-    response.json(blog);
+    response.json(blog)
   } else {
-    response.status(404).end();
+    response.status(404).end()
   }
 })
 
 // POST
 blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body;
+  const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
@@ -36,35 +36,52 @@ blogsRouter.post('/', async (request, response, next) => {
     url: body.url || '',
     likes: body.likes || 0,
     user: user.id
-  });
+  })
 
-  const savedBlog = await blog.save();
+  const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
-  response.status(201).json(savedBlog);
+  response.status(201).json(savedBlog)
 })
 
 // DELETE
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: 'operation not allowed' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+
+  response.status(200).json({ message: 'blog successfully deleted' });
 })
 
 // PUT
 blogsRouter.put('/:id', async (request, response, next) => {
-  const body = request.body;
+  const body = request.body
 
   const blog = {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes
-  };
+  }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
 
-  response.json(updatedBlog);
+  response.json(updatedBlog)
 })
 
 module.exports = blogsRouter
